backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-ratelimit>=4.1.0
transformers>=4.21.0
torch>=2.0.0
numpy>=1.21.0
scikit-learn>=1.1.0
nltk>=3.7
pandas>=1.5.0
requests>=2.28.0
cachetools>=5.2.0
python-dotenv>=0.19.0
ratelimit>=2.2.0
beautifulsoup4>=4.11.0
newspaper3k>=0.2.8
lxml>=4.9.0
lxml_html_clean>=0.4.0
langdetect>=1.0.9